requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
drf-spectacular==0.26.5
django-cors-headers==4.0.0
django-filter==23.1
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
daphne==4.0.0
celery==5.2.7
redis==4.5.4
